backend_test/posture_feedback/requirements.txt
# Posture feedback (standalone test)
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
# For /exercise-feedback-from-image (pose from image)
mediapipe
numpy
pillow
